separated routes from server code
