+++ finalProject/helpers.py
@@ -27,7 +27,7 @@ def login_required(f):
     @wraps(f)
     def decorated_function(*args, **kwargs):
         if session.get("user_id") is None:
-            return redirect(url_for("login", next=request.url))
+            return redirect(url_for("login_api.login", next=request.url))
         return f(*args, **kwargs)
     return decorated_function
