+++ finalProject/templates/layout.html
@@ -38,19 +38,19 @@
                     <div class="collapse navbar-collapse" id="navbar">
                         {% if session.user_id %}
                             <ul class="nav navbar-nav">
-                                <li><a href="{{ url_for('browse') }}">Browse</a></li>
-                                <li><a href="{{ url_for('buy') }}">Buy</a></li>
-                                <li><a href="{{ url_for('sell') }}">Sell</a></li>
-                                <li><a href="{{ url_for('history') }}">User Stories</a></li>
-                                <li><a href="{{ url_for('account') }}">Account</a></li>
+                                <li><a href="{{ url_for('browse_api.browse') }}">Browse</a></li>
+                                <li><a href="{{ url_for('buy_api.buy') }}">Buy</a></li>
+                                <li><a href="{{ url_for('sell_api.sell') }}">Sell</a></li>
+                                <li><a href="{{ url_for('history_api.history') }}">User Stories</a></li>
+                                <li><a href="{{ url_for('account_api.account') }}">Account</a></li>
                             </ul>
                             <ul class="nav navbar-nav navbar-right">
-                                <li><a href="{{ url_for('logout') }}">Log Out</a></li>
+                                <li><a href="{{ url_for('logout_api.logout') }}">Log Out</a></li>
                             </ul>
                         {% else %}
                             <ul class="nav navbar-nav navbar-right">
-                                <li><a href="{{ url_for('register') }}">Register</a></li>
-                                <li><a href="{{ url_for('login') }}">Log In</a></li>
+                                <li><a href="{{ url_for('register_api.register') }}">Register</a></li>
+                                <li><a href="{{ url_for('login_api.login') }}">Log In</a></li>
                             </ul>
                         {% endif %}
                     </div>
