+++ finalProject/templates/login.html
@@ -2,7 +2,7 @@
 
 
 {% block main %}
-    <form action="{{ url_for('login') }}" method="get">
+    <form action="{{ url_for('login_api.login') }}" method="get">
         <fieldset>
             <div class="form-group">
                 <label><b>Username</b></label>
