+++ finalProject/templates/register.html
@@ -5,7 +5,7 @@
 {% endblock %}
 
 {% block main %}
-    <form action="{{ url_for('register') }}" method="post">
+    <form action="{{ url_for('register_api.register') }}" method="post">
         <fieldset>
             <div class="form-group">
                 <input class="form-control" name="name" placeholder="name" type="text"/>
