+++ finalProject/templates/shopping.html
@@ -6,7 +6,7 @@
 
 {% block main %}
 <!DOCTYPE html>
-<form action="{{ url_for('shopping') }}" method="get">
+<form action="{{ url_for('shopping_api.shopping') }}" method="get">
     <fieldset>
         <table style="width:100%">
             <caption>{{ username }} Shopping Cart</caption>
